Refactor BIND and Knot code into their own subclasses
Ideally this is a noop change.
